Refactor the flow of terraform converters, created separate class which encapsulate the logic of adding and composing resources for terraform
