+++ syndicate/core/constants.py
@@ -34,6 +34,7 @@ EC2_INSTANCE_TYPE = 'ec2_instance'
 BATCH_COMPENV_TYPE = 'batch_compenv'
 BATCH_JOBQUEUE_TYPE = 'batch_jobqueue'
 BATCH_JOBDEF_TYPE = 'batch_jobdef'
+DYNAMO_DB_STREAM_TYPE = 'dynamodb_stream'
 
 S3_PATH_NAME = 's3_path'
 
@@ -70,6 +71,7 @@ DEPLOY_RESOURCE_TYPE_PRIORITY = {
     BATCH_COMPENV_TYPE: 20,
     BATCH_JOBQUEUE_TYPE: 21,
     BATCH_JOBDEF_TYPE: 22,
+    DYNAMO_DB_STREAM_TYPE: 23
 }
 
 CLEAN_RESOURCE_TYPE_PRIORITY = {
