+++ syndicate/core/transform/terraform/converter/tf_resource_converter.py
@@ -1,0 +1,15 @@
+from abc import abstractmethod
+
+from syndicate.core.transform.terraform.terraform_template import \
+    TerraformTemplate
+
+
+class TerraformResourceConverter:
+
+    def __init__(self, template: TerraformTemplate, config=None):
+        self.template = template
+        self.config = config
+
+    @abstractmethod
+    def convert(self, name, resource):
+        pass
